Moved the main code into helper allow creation of console tester Added dump logic
